requirements.txt
flet>=0.20.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
urllib3>=2.2.0
PySocks>=1.7.1
loguru